refactor: move functions for change state
