+++ src/components/view/product/Product.ts
@@ -69,13 +69,14 @@ export class ProductPage {
 
     const btn = document.querySelector('#product-add') as HTMLButtonElement;
     btn.onclick = () => {
-      app.controller.setActualState('cart', btn.value, false);
+      app.controller.appStateControl('cart', btn.value, false);
       btn.textContent = this.getCartButtonName();
     };
 
     const buyBtn = document.querySelector('#product-buy') as HTMLButtonElement;
     buyBtn.onclick = () => {
-      if (!STATE.cartProducts.find((item) => item.id === this.product.id)) app.controller.addProdToCart(+buyBtn.value);
+      if (!STATE.cartProducts.find((item) => item.id === this.product.id))
+        app.controller.appStateControl('cart', btn.value, false);
       window.history.replaceState(null, '', window.location.search + '#/cart/');
       app.router.resolveRoute();
       app.view.cart.cartTotal.openPopup();
